+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 1.2.0
 ### Added
 - Added Albedo integration for Colored Redstone Torches (colored light). Colored light for torches can be enabled/disabled in mod config. **Requires [Albedo](https://minecraft.curseforge.com/projects/albedo) mod.**
-- Added Colored Redstone Lamps. 16 new colors. Shade of the lamp and it's light level depends of the power of provided redstone signal.
+- Added Colored Redstone Lamps (16 color variants). Shade of the lamp and it's light level depends of the power of provided redstone signal.
+- Added Albedo integration for Colored Redstone Lamps (colored light). Colored light for lamps can be enabled/disabled in mod config. **Requires [Albedo](https://minecraft.curseforge.com/projects/albedo) mod.**
 ## 1.1.0
 ### Added
 - Added Chisel integration for Colored Redstone Blocks (Redstone group).